fix: resolve dev server crashes for both backend and frontend
- Backend fixes:
  - Remove "type": "module" from package.json to fix ES module errors
  - Convert ESLint config from ES modules to CommonJS
  - Downgrade Express from v5.0.0 to v4.21.2 for compatibility
  - Add transpileOnly to ts-node configuration for faster development
  - Fix database schema missing cost tracking tables
  - Remove duplicate database initialization from CostTracker
  - Fix LRUCache import in ProgressService

- Frontend fixes:
  - Update PostCSS config to use ES module syntax
  - Fix "export default" vs "module.exports" compatibility

Both development servers now start successfully without crashes.
