+++ backend/eslint.config.js
@@ -1,9 +1,9 @@
 // @ts-check
-import eslint from '@eslint/js';
-import tseslint from '@typescript-eslint/eslint-plugin';
-import tsParser from '@typescript-eslint/parser';
+const eslint = require('@eslint/js');
+const tseslint = require('@typescript-eslint/eslint-plugin');
+const tsParser = require('@typescript-eslint/parser');
 
-export default [
+module.exports = [
   // Apply to all TypeScript files
   {
     files: ['**/*.ts', '**/*.tsx'],
