+++ backend/package.json
@@ -2,7 +2,6 @@
   "name": "prompt-card-backend",
   "version": "1.0.0",
   "private": true,
-  "type": "module",
   "scripts": {
     "dev": "nodemon src/server.ts",
     "build": "tsc",
@@ -25,7 +24,7 @@
     "csv-writer": "^1.6.0",
     "dotenv": "^17.0.0",
     "exceljs": "^4.4.0",
-    "express": "^5.0.0",
+    "express": "^4.21.2",
     "helmet": "^8.0.0",
     "html-pdf-node": "^1.0.7",
     "ioredis": "^5.6.1",
