+++ backend/tsconfig.json
@@ -6,6 +6,11 @@
     "outDir": "./dist",
     "rootDir": "./src",
     "strict": false,
+    "noImplicitAny": false,
+    "strictNullChecks": false,
+    "strictFunctionTypes": false,
+    "strictBindCallApply": false,
+    "strictPropertyInitialization": false,
     "esModuleInterop": true,
     "skipLibCheck": true,
     "noImplicitAny": false,
@@ -26,6 +31,15 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["src/*"]
+    },
+    "allowSyntheticDefaultImports": true
+  },
+  "ts-node": {
+    "transpileOnly": true,
+    "compilerOptions": {
+      "strict": false,
+      "noImplicitAny": false,
+      "strictNullChecks": false
     }
   },
   "include": [
